[datafeeder] remove temp files once publish job is complete

+++ datafeeder/src/main/java/org/georchestra/datafeeder/service/DataUploadService.java
@@ -25,6 +25,8 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+import javax.transaction.Transactional;
+
 import org.georchestra.datafeeder.api.FileUploadApiController;
 import org.georchestra.datafeeder.batch.service.DataUploadAnalysisService;
 import org.georchestra.datafeeder.model.BoundingBoxMetadata;
@@ -47,6 +49,7 @@ public class DataUploadService {
     private @Autowired DataUploadJobRepository repository;
     private @Autowired DataUploadAnalysisService analysisService;
     private @Autowired DatasetsService datasetsService;
+    private @Autowired FileStorageService storageService;
 
     public Optional<DataUploadJob> findJob(UUID uploadId) {
         return repository.findByJobId(uploadId);
@@ -79,8 +82,23 @@ public class DataUploadService {
         throw new UnsupportedOperationException("unimplemented");
     }
 
+    @Transactional
     public void remove(@NonNull UUID jobId) {
-        repository.deleteById(jobId);
+        Optional<DataUploadJob> found = findJob(jobId);
+        if (found.isPresent()) {
+            DataUploadJob job = found.get();
+            if (JobStatus.RUNNING == job.getAnalyzeStatus()) {
+                throw new IllegalStateException("Can't remove a job while the analysis process is running");
+            }
+            if (JobStatus.RUNNING == job.getPublishStatus()) {
+                throw new IllegalStateException("Can't remove a job while the publish process is running");
+            }
+            try {
+                repository.deleteById(jobId);
+            } finally {
+                this.storageService.deletePackage(jobId);
+            }
+        }
     }
 
     public SimpleFeature sampleFeature(@NonNull UUID jobId, @NonNull String typeName, int featureN, Charset encoding,
